# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cloud Assessment MVP API
  version: 0.1.0
  description: |
    AWS理解度テストWebアプリMVPのAPI定義。
    認証は署名付きHTTP-only Cookie（session）を利用する。
servers:
  - url: http://localhost:3000
paths:
...
  /api/attempts/create:
    post:
      tags: [Attempts]
      summary: Attempt作成
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttemptRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAttemptResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/attempts/{attemptId}/answer:
    post:
      tags: [Attempts]
      summary: 回答保存（1問につき1回のみ）
      security:
        - cookieAuth: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerAttemptRequest"
      responses:
        "200":
          description: 保存成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerAttemptResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/attempts/{attemptId}:
    get:
      tags: [Attempts]
      summary: Attempt詳細取得
      description: 指定Attemptの問題一覧・進捗・結果を返す。/quiz/[attemptId]画面で使用。
      security:
        - cookieAuth: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/me/attempts/{attemptId}:
    get:
      tags: [Me]
      summary: 受験結果詳細取得
      description: 特定Attemptの詳細結果（間違い問題・解説含む）を返す。
      security:
        - cookieAuth: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyAttemptDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  responses:
    BadRequest:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未認証
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 権限不足（所有者不一致など）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 対象が存在しない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    CreateAttemptRequest:
      type: object
      required: [categories, level, numQuestions]
      properties:
        categories:
          type: array
          items:
            type: string
            maxLength: 200
          minItems: 1
          maxItems: 100
        level:
          type: integer
          enum: [1, 2, 3]
        numQuestions:
          type: integer
          minimum: 1
    CreateAttemptResponse:
      type: object
      required: [attemptId]
      properties:
        attemptId:
          type: string
    AnswerAttemptRequest:
      type: object
      required: [attemptQuestionId, selectedIndex]
      properties:
        attemptQuestionId:
          type: string
          minLength: 1
          maxLength: 100
        selectedIndex:
          type: integer
          minimum: 0
          maximum: 3
    AnswerAttemptResponse:
      type: object
      required: [attemptQuestionId, selectedIndex]
      properties:
        attemptQuestionId:
          type: string
        selectedIndex:
          type: integer
    CategoryBreakdownItem:
      type: object
      required: [correct, total, percent]
      properties:
        correct:
          type: integer
        total:
          type: integer
        percent:
          type: number
          format: float
    FinalizeAttemptResponse:
      type: object
      required: [overallPercent, categoryBreakdown]
      properties:
        overallPercent:
          type: number
          format: float
        categoryBreakdown:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/CategoryBreakdownItem"
...
    AttemptQuestionItem:
      type: object
      required: [id, questionId, order, questionText, choices]
      properties:
        id:
          type: string
        questionId:
          type: string
        order:
          type: integer
        questionText:
          type: string
        choices:
          type: array
          items:
            type: string
        selectedIndex:
          type: integer
          nullable: true
        isCorrect:
          type: boolean
          nullable: true
          description: COMPLETED時のみ非null（IN_PROGRESS中はnull）
        answerIndex:
          type: integer
          description: COMPLETED時のみ返却
        explanation:
          type: string
          description: COMPLETED時のみ返却
        category:
          type: string
    AttemptDetailResponse:
      type: object
      required: [id, status, filters, startedAt, questions]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [IN_PROGRESS, COMPLETED]
        filters:
          type: object
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        questions:
          type: array
          items:
            $ref: "#/components/schemas/AttemptQuestionItem"
        result:
          $ref: "#/components/schemas/FinalizeAttemptResponse"
          nullable: true
...
    MyAttemptDetailResponse:
      type: object
      required: [id, status, filters, startedAt, questions, result]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [IN_PROGRESS, COMPLETED]
        filters:
          type: object
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        questions:
          type: array
          items:
            $ref: "#/components/schemas/AttemptQuestionItem"
        result:
          $ref: "#/components/schemas/FinalizeAttemptResponse"
...